refactor(Tests Knowledge): 拆分 word_boundary 测试、合并 graph_service 碎片、清理 pipeline_tracker 重复 Fakes;
- test_chunking.py 拆出 TestWordBoundaryProtection/EdgeCases → test_chunking_word_boundary.py
- test_graph_service_model_name.py（46行）合入 test_graph_service.py，删除碎片文件
- test_pipeline_tracker.py 移除 4 个重复 Fake 类定义，改从 conftest 导入（-56行）
